+++ homeassistant/components/fritzbox/binary_sensor.py
@@ -1,17 +1,49 @@
 """Support for Fritzbox binary sensors."""
 from __future__ import annotations
 
+from dataclasses import dataclass
+from typing import Callable, Final
+
 from pyfritzhome.fritzhomedevice import FritzhomeDevice
 
-from homeassistant.components.binary_sensor import BinarySensorEntity
-from homeassistant.components.fritzbox.model import FritzBinarySensorEntityDescription
+from homeassistant.components.binary_sensor import (
+    DEVICE_CLASS_WINDOW,
+    BinarySensorEntity,
+    BinarySensorEntityDescription,
+)
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from . import FritzBoxEntity
-from .const import BINARY_SENSOR_TYPES, CONF_COORDINATOR, DOMAIN as FRITZBOX_DOMAIN
+from .const import CONF_COORDINATOR, DOMAIN as FRITZBOX_DOMAIN
+from .model import FritzEntityDescriptionMixinBase
+
+
+@dataclass
+class FritzEntityDescriptionMixinBinarySensor(FritzEntityDescriptionMixinBase):
+    """BinarySensor description mixin for Fritz!Smarthome entities."""
+
+    is_on: Callable[[FritzhomeDevice], bool | None]
+
+
+@dataclass
+class FritzBinarySensorEntityDescription(
+    BinarySensorEntityDescription, FritzEntityDescriptionMixinBinarySensor
+):
+    """Description for Fritz!Smarthome binary sensor entities."""
+
+
+BINARY_SENSOR_TYPES: Final[tuple[FritzBinarySensorEntityDescription, ...]] = (
+    FritzBinarySensorEntityDescription(
+        key="alarm",
+        name="Alarm",
+        device_class=DEVICE_CLASS_WINDOW,
+        suitable=lambda device: device.has_alarm,  # type: ignore[no-any-return]
+        is_on=lambda device: device.alert_state,  # type: ignore[no-any-return]
+    ),
+)
 
 
 async def async_setup_entry(
